added react-icons to project. Refactored index.js file so now it takes our express app and listens on specified port.
