+++ client/package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.4.0",
     "react-loader-spinner": "^6.0.0-0",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
